Made fixes as per second review(Commented Auth, used Neon.tech for db and updated the required files)
