+++ .env.example
@@ -1,4 +1,5 @@
-DATABASE_URL="postgresql://postgres:123456@localhost:5432/postgres"
+# DATABASE_URL="postgresql://postgres:123456@localhost:5432/postgres"
+DATABASE_URL="postgresql://neondb_owner:npg_9pAFuH8eKLzJ@ep-blue-wave-a8isq30a-pooler.eastus2.azure.neon.tech/neondb?sslmode=require"
 
 # JWT Secrets by role
 JWT_SECRET_USER=""
