+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/index.ts","./src/graphql/index.ts","./src/graphql/user/index.ts","./src/graphql/user/mutations.ts","./src/graphql/user/queries.ts","./src/graphql/user/resolver.ts","./src/graphql/user/typedefs.ts","./src/lib/db.ts","./src/services/user.ts"],"version":"5.8.2"}
+{"root":["./src/index.ts","./src/graphql/index.ts","./src/graphql/address/userresolvers.ts","./src/graphql/address/usertypes.ts","./src/graphql/user/index.ts","./src/graphql/user/usermutations.ts","./src/graphql/user/userqueries.ts","./src/graphql/user/userresolver.ts","./src/graphql/user/usertypedefs.ts","./src/lib/db.ts","./src/services/address.ts","./src/services/user.ts"],"version":"5.8.2"}
